requirements.txt
mysql.connector
serial
datetime
requests
flake8
python-dotenv